+++ tests/Fixtures/UnitEnums/Makers/MakersUnitEnum.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Henzeb\Enumhancer\Tests\Fixtures\UnitEnums\Makers;
+
+use Henzeb\Enumhancer\Concerns\From;
+use Henzeb\Enumhancer\Concerns\Makers;
+
+enum MakersUnitEnum
+{
+    use Makers, From;
+
+    case Zero;
+    case One;
+    case Three;
+}